scripts/rl/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
apache-beam==2.48.0
apache-flink==1.20.0
apache-flink-libraries==1.20.0
attrs==24.2.0
avro-python3==1.10.2
black==24.8.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
colorful==0.5.6
contourpy==1.3.1
crcmod==1.7
cycler==0.12.1
dill==0.3.1.1
distlib==0.3.9
dm-tree==0.1.8
dnspython==2.7.0
docopt==0.6.2
Farama-Notifications==0.0.4
fastapi==0.115.5
fastavro==1.9.7
fasteners==0.19
filelock==3.16.1
find_libpython==0.4.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.67.1
gymnasium==1.0.0
h11==0.14.0
hdfs==2.7.3
httplib2==0.22.0
httptools==0.6.4
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.43.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.14.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgpack-numpy==0.4.8
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
objsize==0.6.1
opencensus==0.11.4
opencensus-context==0.1.3
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pemja==0.4.1
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.23.4
py-spy==0.4.0
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
Pygments==2.18.0
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
starlette==0.41.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
textual==0.85.2
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.5.1
tqdm==4.67.0
triton==3.1.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.27.1
watchfiles==0.24.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.16.0
yarl==1.17.1
zstandard==0.23.0
